==> Part 1.ipynb <==
# Import Libraries

import pandas as pd
import numpy as np

# Load Data

title_basics = 'https://datasets.imdbws.com/title.basics.tsv.gz'
title_akas = 'https://datasets.imdbws.com/title.akas.tsv.gz'
title_ratings = 'https://datasets.imdbws.com/title.ratings.tsv.gz'

basics = pd.read_csv(title_basics, sep='\t', low_memory=False)
akas = pd.read_csv(title_akas, sep='\t', low_memory=False)
ratings = pd.read_csv(title_ratings, sep='\t', low_memory=False)

# AKAs 

I will start the clean up process on the AKAs table first so that I can use it to filter other dataframes for the specification of keeping only US movies.

# Display the column names of akas
akas.info()

### Replace "\N" with np.nan

# Replace "\N" with np.nan
akas.replace({'\\N':np.nan}, inplace=True)

### Keep only US movies

# Display only US movies
filter_US = akas['region'] == 'US'
akas[filter_US]

# Keep only US movies
akas = akas[filter_US]

### Cleaned AKAs Dataframe

# Cleaned akas table per specification
akas.info()

# Basics

# Display the column names of basics
basics.info()

### Replace "\N" with np.nan

# Replace "\N" with np.nan
basics.replace({'\\N':np.nan}, inplace=True)

### Eliminate movies that are null for runtimeMinutes

#Display the rows that contain null values in the runtimeMinutes column
filter1 = basics['runtimeMinutes'].isna() == True
basics[filter1]

# Eliminate movies that are null for runtimeMinutes
basics = basics[~filter1]

### Eliminate movies that are null for genre

# Display the rows that contain null values in the genres column
filter2 = basics['genres'].isna() == True
basics[filter2]

# Eliminate movies that are null for runtimeMinutes
basics = basics[~filter2]

### Keep only titleType==Movie

# Display only titleType==Movie
filter3 = basics['titleType'] == 'Movie'
basics[filter3]

basics['titleType'].value_counts()

# Display only titleType==Movie
filter3 = basics['titleType'] == 'movie'
basics[filter3]

# Keep only titleType==Movie
basics = basics[filter3]

###  Keep startYear 2000 - 2021

# Convert dtype of startYear column to float
basics['startYear'].astype(float)

# Filter for startYear 2000-2021
filter4 = basics['startYear'] >= '2000'
filter5 = basics['startYear'] <= '2021'
basics[filter4 & filter5]

# Keep only startYear 2000-2021
basics = basics[filter4 & filter5]

### Eliminate movies that include "Documentary" in genre

# Display movies that are included in the documentary category.
is_documentary = basics['genres'].str.contains('documentary',case=False)
basics[is_documentary]

# Exclude movies that are included in the documentary category.
basics = basics[~is_documentary]

### Keep only US movies using AKAs table

# Keep only US movies
# Filter the basics table down to only include the US by using the filter akas dataframe
keepers =basics['tconst'].isin(akas['titleId'])
basics = basics[keepers]

### Cleaned Basics Dataframe

# Cleaned Basics table per specification
basics.info()

# Ratings

# Display the column names of akas
ratings.info()

### Replace "\N" with np.nan

# Replace "\N" with np.nan
ratings.replace({'\\N':np.nan}, inplace=True)

### Keep only US movies using AKAs table

# Keep only US movies
# Filter the ratings table down to only include the US by using the filter akas dataframe
keepers2 =ratings['tconst'].isin(akas['titleId'])
ratings = ratings[keepers2]

### Cleaned Ratings Dataframe

# Cleaned Ratings table per specification
ratings.info()

# Final .info() of each dataframe

print('Basics', '\n')
print(basics.info())
print('\n','AKAs', '\n')
print(akas.info())
print('\n','Ratings', '\n')
print(ratings.info())

# Create a "Data" folder

import os
os.makedirs('Data/',exist_ok=True) 
# Confirm folder created
os.listdir("Data/")

# Save each file to a compressed csv file "Data/" folder inside your repository.

## Save current dataframe to file.
basics.to_csv("Data/title_basics.csv.gz",compression='gzip',index=False)

akas.to_csv("Data/title_akas.csv.gz",compression='gzip',index=False)

ratings.to_csv("Data/title_ratings.csv.gz",compression='gzip',index=False)

# Check the saved csv file

# Open saved file and preview again
basics = pd.read_csv("Data/title_basics.csv.gz", low_memory = False)
basics.head()

# Open saved file and preview again
akas = pd.read_csv("Data/title_akas.csv.gz", low_memory = False)
akas.head()

# Open saved file and preview again
ratings = pd.read_csv("Data/title_ratings.csv.gz", low_memory = False)
ratings.head()
