==> Part 3.ipynb <==
# 1. Import Library

import pandas as pd

# 2. Load Data

## Title Basics

basics = pd.read_csv('Data/title_basics.csv.gz')
basics.head()

## Title Ratings

ratings = pd.read_csv('Data/title_ratings.csv.gz')
ratings.head()

## The TMDB API Results (tmdb_results_combined)

tmdb = pd.read_csv('Data/tmdb_results_combined.csv.gz')
tmdb.head()

# 3. Inspect and Prepare Data

basics.info()

ratings.info()

tmdb.info()

## Required Transformation Steps for Title Basics:
#### Normalize Genre:

- Convert the single string of genres from title basics into 2 new tables.
1. title_genres: with the columns:

   - tconst
   - genre_id


2. genres:
   - genre_id
   - genre_name
   
#### Discard unnecessary information:

- For the `title basics` table, drop the following columns:
    - "original_title" (we will use the primary title column instead)
    - "isAdult" ("Adult" will show up in the genres so this is redundant information).
    - "titleType" (every row will be a movie).
    - "genres" and other variants of genre (genre is now represented in the 2 new tables described above.

- Do not include the `title_akas` table in your SQL database.
    - You have already filtered out the desired movies using this table, and the remaining data is mostly nulls and not of interest to the stakeholder.

## 3.1 Normalize Genre

##### 1. Get a list of all individual genres.

- First, I will need to get a list of all of the unique genres that appear in the column. 
- Right now, the genre column contains a string with the genres separated by a comma.
- I am going to convert these strings into lists of strings into a new 'genres_split' column.

# check the current column names
basics.info()

## create a col with a list of genres
basics['genres_split'] = basics['genres'].str.split(',')
basics

**Use .explode() to separate the list of genres into new rows: one row for each genre a movie belonged to.**

# explode the series using .explode()
exploded_genres = basics.explode('genres_split')
exploded_genres

**Use .unique() to get the unique genres from the genres_split column and sort alphabetically**

unique_genres = sorted(exploded_genres['genres_split'].unique())

unique_genres

##### 2. Create a new `title_genres` table

We only want the tconst and genres_split columns.

#Create a new dataframe which only contains the necessary columns
title_genres = exploded_genres[['tconst', 'genres_split']].copy()
title_genres.head()

##### 3. Create a genre mapper dictionary to replace string genres with integers

## Making the genre mapper dictionary
### Make a dictionary with list of unique genres as the key and unique integer ids as values
genre_ints = range(len(unique_genres))
genre_map = dict(zip(unique_genres, genre_ints))
genre_map

##### 4. Replace the string genres in title_genres with the new integer ids.

## make a new integer column called "genre_id", and drop the string "genres_split" column
title_genres['genre_id'] = title_genres['genres_split'].map(genre_map)
title_genres = title_genres.drop(columns='genres_split')
title_genres.head()

Now the `title_genres` table is ready for MySQL database.

##### 5. Convert the genre map dictionary into a dataframe.

# Make a dataframe with Genre_Name and Genre_ID columns
genre = pd.DataFrame({'Genre_Name': genre_map.keys(), 
                      'Genre_ID': genre_map.values()})
genre.head()

Now the `genres` table is ready for MySQL database.

## 3.2 Saving the MySQL tables with tconst as the primary key.
