==> Part 4 - API calls to obtain more data.ipynb <==
# Import Library

# Standard Imports
import numpy as np
import pandas as pd

# Additional Imports
import os, json, math, time
import tmdbsimple as tmdb
from tqdm.notebook import tqdm_notebook

# Load Credentials and Create Movie Object

import json
with open('/Users/ericakitano/.secret/tmdb_api.json', 'r') as f:
    login = json.load(f)
## Display the keys of the loaded dict
login.keys()

# Log-in to TMDB using my API-key
tmdb.API_KEY =  login['API-key']

## make a movie object using the .Movies function from tmdb
movie = tmdb.Movies(603)

# Obtain Certification Data from `releases()`

## Define `get_movie_with_rating` function

def get_movie_with_rating(movie_id):
    
    # Get the movie object for the current id
    movie = tmdb.Movies(movie_id)
    
    # save the .info .releases dictionaries
    info=movie.info()
    
    releases = movie.releases()
    
    #Loop through countries in releases
    for c in releases['countries']:
        #if the country is US
        if c['iso_3166_1'] == 'US':
            # save a "certification" key in "info" with the certification
            info['certification'] = c['certification']
    return info


## Test `get_movie_with_rating` function

### Test#1: movie_id: tt0848228 ("The Avengers")

# Test get_movie_with_rating function with movie_id: tt0848228 ("The Avengers")
test = get_movie_with_rating("tt0848228")
test

Verified that `info` now contains `certification` key. 

### Test#2 : movie_id: tt0332280 ("The Notebook")

# Test get_movie_with_rating function with movie_id: tt0332280 ("The Notebook")
test2 = get_movie_with_rating("tt0332280")
test2

Verified that the `get_movie_with_rating` function is working.

# Obtain Financial Data

### Designate a folder to save data obtained from API calls

FOLDER = "Data/"
os.makedirs(FOLDER, exist_ok=True)
os.listdir(FOLDER)

### Define `write_json` function

This function appends a list of records to a json file.

def write_json(new_data, filename): 
    """Appends a list of records (new_data) to a json file (filename). 
    Adapted from: https://www.geeksforgeeks.org/append-to-json-file-using-python/"""  
    
    with open(filename,'r+') as file:
        # First we load existing data into a dict.
        file_data = json.load(file)
        ## Choose extend or append
        if (type(new_data) == list) & (type(file_data) == list):
            file_data.extend(new_data)
        else:
             file_data.append(new_data)
        # Sets file's current position at offset.
        file.seek(0)
        # convert back to json.
        json.dump(file_data, file)

### Load in the Title Basics data

# Load in the dataframe from project part 1 as basics:
basics = pd.read_csv("Data/title_basics.csv.gz")

### Define a list of the Years to Extract from the API

Create Required Lists for the Outer Loop

YEARS_TO_GET = range(2002,2022)

### Define an errors list

Create an empty errors list before the loops so that we can append and save the ids and error messages for any movie that causes an error. 

errors = [ ]

### Nested For Loops to obtain data and store

# Start of OUTER loop

## Set up Progress Bar which will iterate through each year that is defined in the YEARS_TO_GET variable.
for YEAR in tqdm_notebook(YEARS_TO_GET, desc='YEARS', position=0):
    
    #Defining the JSON file to store results for year
    JSON_FILE = f'{FOLDER}tmdb_api_results_{YEAR}.json'
    
    # Check if file exists
    file_exists = os.path.isfile(JSON_FILE)
    
    # If it does not exist: create it
    if file_exists == False:
        # save an empty dict with just "imdb_id" to the new json file.
        with open(JSON_FILE,'w') as f:
            json.dump([{'imdb_id':0}],f)
            
    #Saving new year as the current df
    df = basics.loc[ basics['startYear']==YEAR].copy()
    # saving movie ids to list
    movie_ids = df['tconst'].copy()
            
    # Load existing data from json into a dataframe called "previous_df"
    previous_df = pd.read_json(JSON_FILE)
            
    # filter out any ids that are already in the JSON_FILE
    movie_ids_to_get = movie_ids[~movie_ids.isin(previous_df['imdb_id'])]
            
    ## Start INNER Loop
    
    #Get index and movie id from list
    for movie_id in tqdm_notebook(movie_ids_to_get,
                                  desc=f'Movies from {YEAR}',
                                  position=1,
                                  leave=True):
        try:
            # Retrieve then data for the movie id
            temp = get_movie_with_rating(movie_id)  
            # Append/extend results to existing file using a pre-made function
            write_json(temp,JSON_FILE)
            # Short 20 ms sleep to prevent overwhelming server
            time.sleep(0.02)
            
        except Exception as e:
            errors.append([movie_id, e])
            
        # Save the year's results as csv.gz file
        final_year_df = pd.read_json(JSON_FILE)
        final_year_df.to_csv(f"{FOLDER}final_tmdb_data_{YEAR}.csv.gz", compression="gzip", index=False)

# Print a message reporting back the number of movie ids that caused an error.
print(f"- Total errors: {len(errors)}")

# Check the retrieved data

final_tmdb_data_2002 = pd.read_csv('Data/final_tmdb_data_2002.csv.gz')
final_tmdb_data_2002.head(2)

# Merge all data

import glob
q = "Data/final_tmdb_data_*.csv.gz"
chunked_files = sorted(glob.glob(q))
# Show up to first 30 files
chunked_files[:30]

## Loading all files as df and appending to a list
df_list = []
for file in chunked_files:
    temp_df = pd.read_csv(file, index_col=0, lineterminator='\n')
    df_list.append(temp_df)
    
## Concatenating the list of dfs into 1 combined
df_combined = pd.concat(df_list)
df_combined.head(2)

# Save a final merged .csv.gz of all of the tmdb api data

## Save current dataframe to file.
df_combined.to_csv("Data/part4_tmdb_results_all.csv.gz",compression='gzip',index=False)
